+++ sphere.cpp
@@ -1,8 +1,8 @@
 #include "sphere.h"
 
-bool sphere_intersect(Sphere spheres[], int n, const Ray & r, int &id, double &t)
+bool sphere_intersect(Sphere spheres[], int n, const Ray & r, int &id, float &t)
 {
-    double d;
+    float d;
     t = 1e20;
     for (int i = int (n); i--;)
         if ((d = spheres[i].intersect(r)) && d < t)
